backend-django/requirements.txt
Django
djangorestframework
django-cors-headers
pandas
reportlab
matplotlib
PyQt5
